add get server pool
