+++ README.md
@@ -24,33 +24,38 @@ import "github.com/silenceper/wechat/v2"
 以下是一个微信公众号处理消息接收以及回复的例子：
 
 ```go
-//使用memcache保存access_token，也可选择redis或自定义cache
+// 使用memcache保存access_token，也可选择redis或自定义cache
 wc := wechat.NewWechat()
 memory := cache.NewMemory()
 cfg := &offConfig.Config{
-    AppID:     "xxx",
-    AppSecret: "xxx",
-    Token:     "xxx",
-    //EncodingAESKey: "xxxx",
-    Cache: memory,
+AppID:     "xxx",
+AppSecret: "xxx",
+Token:     "xxx",
+//EncodingAESKey: "xxxx",
+Cache: memory,
 }
 officialAccount := wc.GetOfficialAccount(cfg)
 
 // 传入request和responseWriter
 server := officialAccount.GetServer(req, rw)
-//设置接收消息的处理方法
-server.SetMessageHandler(func(msg *message.MixMessage) *message.Reply {
 
-    //回复消息：演示回复用户发送的消息
-    text := message.NewText(msg.Content)
-    return &message.Reply{MsgType: message.MsgTypeText, MsgData: text}
+// 采用 对象复用的请使用下面这两个方法。
+// server := officialAccount.GetServerByPool(req, rw)
+// defer server.Close()
+
+//设置接收消息的处理方法
+server.SetMessageHandler(func (msg *message.MixMessage) *message.Reply {
+
+//回复消息：演示回复用户发送的消息
+text := message.NewText(msg.Content)
+return &message.Reply{MsgType: message.MsgTypeText, MsgData: text}
 })
 
 //处理消息接收以及回复
 err := server.Serve()
 if err != nil {
-    fmt.Println(err)
-    return
+fmt.Println(err)
+return
 }
 //发送回复的消息
 server.Send()
