+++ officialaccount/server/util.go
@@ -16,7 +16,7 @@ func writeContextType(w http.ResponseWriter, value []string) {
 }
 
 // Render render from bytes
-func (srv *Server) Render(bytes []byte) {
+func (srv *OfficialServer) Render(bytes []byte) {
 	// debug
 	// fmt.Println("response msg = ", string(bytes))
 	srv.Writer.WriteHeader(200)
@@ -27,13 +27,13 @@ func (srv *Server) Render(bytes []byte) {
 }
 
 // String render from string
-func (srv *Server) String(str string) {
+func (srv *OfficialServer) String(str string) {
 	writeContextType(srv.Writer, plainContentType)
 	srv.Render([]byte(str))
 }
 
 // XML render to xml
-func (srv *Server) XML(obj interface{}) {
+func (srv *OfficialServer) XML(obj interface{}) {
 	writeContextType(srv.Writer, xmlContentType)
 	bytes, err := xml.Marshal(obj)
 	if err != nil {
@@ -43,16 +43,19 @@ func (srv *Server) XML(obj interface{}) {
 }
 
 // Query returns the keyed url query value if it exists
-func (srv *Server) Query(key string) string {
+func (srv *OfficialServer) Query(key string) string {
 	value, _ := srv.GetQuery(key)
 	return value
 }
 
 // GetQuery is like Query(), it returns the keyed url query value
-func (srv *Server) GetQuery(key string) (string, bool) {
+func (srv *OfficialServer) GetQuery(key string) (string, bool) {
 	req := srv.Request
 	if values, ok := req.URL.Query()[key]; ok && len(values) > 0 {
 		return values[0], true
 	}
 	return "", false
 }
+
+// Close this server
+func (srv *OfficialServer) Close() {}
